requirements.txt
flask
flask-cors
python-dotenv
langgraph
pydantic
langchain-google-genai
langchain-core
